Refactor to minimize changes to syn, move tests to compile.rs
